# [WIP] Patch for CLIP