@dbolya 2023-03-12
Hi thanks for trying this out. Yeah the native timm implementation for Attention is not very efficient.

There are two things we need from attn:
1. The key values (k) to use as metrics for merging
2. To add sizes to the pre (or post) softmax attn matrix

The problem is, as you've discovered, is that getting / setting those values from `nn.MultiHeadAttn` is very much non-trivial, since it computes K and applies the softmax internally.

For 1, we have the `in_proj_weight` matrix (which includes `W_k`), so we could just re-compute the keys ourselves (which would take extra time, but it's not the end of the world). For 2, there may be a hacky way to add what we want to the pre-softmax matrix by modifying the biases (i.e., `in_proj_bias` or `k_bias`, idk which) on the fly.

Neither 1 or 2 are strict requirements, though, but of course they make the result more accurate.

First, I would try what happens if you use `nn.MultiHeadAttn` without 1 or 2. Simply call the normal attention module (not ToMeAttn) and then use the input `x` as the metric (instead of `k`). Once that gives a speed improvement, then we can start thinking of making it more accurate by incorporating 1 or 2.