@AlexKoff88 2023-05-09
Hi, I used this implementation partially fixed some issues and addressed performance problem that comes from an unnecessary permutation of tensor dimensions. I applied it to OpenCLIP models. Please find my implementation [here](https://github.com/AlexKoff88/tomesd/blob/main/tomeov/openclip.py). Any feedback is appreciated.